Changed makefile and added make install/uninstall rule
