+++ Makefile
@@ -3,6 +3,7 @@ TARGETS = \
 	feature \
 	libsvm \
 	wrapper
+  
 
 all:
 	-for dir in $(TARGETS); do \
@@ -18,5 +19,26 @@ clean:
 
 test:
 	@echo hello;
-
+ 
 .PHONY: clean $(TARGETS)
+
+
+PREFIX = /usr
+
+alib = libvmaf.a
+obj = $(wildcard wrapper/obj/*.o)
+
+$(alib): $(obj)
+	ar rcs $@ $^
+
+.PHONY: install
+install: $(alib)
+	mkdir -p $(DESTDIR)$(PREFIX)/lib
+	mkdir -p $(DESTDIR)$(PREFIX)/include
+	cp $(alib) $(DESTDIR)$(PREFIX)/lib/$(alib)
+	cp wrapper/src/vmaf.h $(DESTDIR)$(PREFIX)/include/
+
+.PHONY: uninstall
+uninstall:
+	rm -f $(DESTDIR)$(PREFIX)/lib/$(alib)
+	rm -f $(DESTDIR)$(PREFIX)/include/vmaf.h
