+++ wrapper/src/main.cpp
@@ -21,7 +21,6 @@
 #include <string>
 #include <algorithm>
 
-#include "main.h"
 #include "cpu.h"
 #include "vmaf.h"
 
@@ -47,35 +46,6 @@ void print_usage(int argc, char *argv[])
     fprintf(stderr, "Usage: %s fmt width height ref_path dis_path model_path [--log log_path] [--log-fmt log_fmt] [--disable-clip] [--disable-avx] [--psnr] [--ssim] [--ms-ssim] [--phone-model]\n", argv[0]);
     fprintf(stderr, "fmt:\n\tyuv420p\n\tyuv422p\n\tyuv444p\n\tyuv420p10le\n\tyuv422p10le\n\tyuv444p10le\n\n");
     fprintf(stderr, "log_fmt:\n\txml (default)\n\tjson\n\n");
-}
-
-double compute_vmaf(char* fmt, int width, int height, char *ref_path, char *dis_path)
-{
-    char *model_path;
-    char *log_path = NULL;
-    char *log_fmt = NULL;
-    bool disable_clip = false;
-    bool disable_avx = false;
-    bool enable_transform = false;
-    bool do_psnr = false;
-    bool do_ssim = false;
-    bool do_ms_ssim = false;
-    char *pool_method = NULL;
-
-
-    try
-    {
-        score = RunVmaf(fmt, width, height, ref_path, dis_path, model_path, log_path, log_fmt, disable_clip, enable_transform, do_psnr, do_ssim, do_ms_ssim, pool_method);
-
-        return score;
-    }
-    catch (const std::exception &e)
-    {
-        fprintf(stderr, "Error: %s\n", e.what());
-        print_usage(argc, argv);
-        return -1;
-    }
-
 }
 
 int main(int argc, char *argv[])
