--- wrapper/src/main.h
@@ -1,1 +1,0 @@
-double compute_vmaf(char* fmt, int width, int height, char *ref_path, char *dis_path);
